refactor: 예약 관련 검증로직들 -> reservations패키지 내부로 이동
