+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -1,10 +1,14 @@
-package roomescape.global.error;
+package roomescape.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.reservations.exception.DuplicateReservationException;
+import roomescape.reservations.exception.InvalidReservationTimeException;
+import roomescape.reservations.exception.ReservationFullException;
+import roomescape.reservations.exception.ReservationNotFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
