+++ src/main/java/roomescape/reservations/exception/ReservationFullException.java
@@ -1,4 +1,4 @@
-package roomescape.global.error;
+package roomescape.reservations.exception;
 
 public class ReservationFullException extends RuntimeException {
